chore: move fixtures to single package
