+++ .changeset/config.json
@@ -7,9 +7,6 @@
   "baseBranch": "master",
   "updateInternalDependencies": "patch",
   "ignore": [
-    "@api-viewer-fixtures/expansion-panel",
-    "@api-viewer-fixtures/fancy-accordion",
-    "@api-viewer-fixtures/intl-currency",
-    "@api-viewer-fixtures/progress-bar"
+    "@api-viewer-fixtures/lit"
   ]
 }
