+++ .changeset/pre.json
@@ -7,10 +7,7 @@
     "@api-viewer/docs": "1.0.0-pre.0",
     "@api-viewer/tabs": "1.0.0-pre.0",
     "api-viewer-element": "1.0.0-pre.0",
-    "@api-viewer-fixtures/expansion-panel": "0.0.0",
-    "@api-viewer-fixtures/fancy-accordion": "0.0.0",
-    "@api-viewer-fixtures/intl-currency": "0.0.0",
-    "@api-viewer-fixtures/progress-bar": "0.0.0"
+    "@api-viewer-fixtures/lit": "0.0.0"
   },
   "changesets": []
 }
