--- fixtures/fancy-accordion/tsconfig.json
@@ -1,22 +1,0 @@
-// Don't edit this file directly. It is generated by /scripts/update-package-configs.ts
-
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "module": "ESNext",
-    "outDir": "./lib",
-    "rootDir": "./src",
-    "composite": true
-  },
-  "references": [
-    {
-      "path": "../expansion-panel/tsconfig.json"
-    }
-  ],
-  "include": [
-    "src"
-  ],
-  "exclude": [
-    "lib"
-  ]
-}
