--- fixtures/intl-currency/tsconfig.json
@@ -1,18 +1,0 @@
-// Don't edit this file directly. It is generated by /scripts/update-package-configs.ts
-
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "module": "ESNext",
-    "outDir": "./lib",
-    "rootDir": "./src",
-    "composite": true
-  },
-  "references": [],
-  "include": [
-    "src"
-  ],
-  "exclude": [
-    "lib"
-  ]
-}
