+++ fixtures/lit/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@api-viewer-fixtures/lit",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "analyze": "cem analyze --litelement --globs 'src/*.ts'"
+  },
+  "dependencies": {
+    "lit": "^2.0.0",
+    "tslib": "^2.3.1"
+  },
+  "customElements": "custom-elements.json"
+}
