+++ fixtures/lit/src/expansion-panel.ts
