+++ fixtures/lit/src/intl-currency.ts
