+++ fixtures/lit/src/progress-bar.ts
