--- fixtures/progress-bar/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@api-viewer-fixtures/progress-bar",
-  "version": "0.0.0",
-  "private": true,
-  "main": "lib/progress-bar.js",
-  "module": "lib/progress-bar.js",
-  "dependencies": {
-    "lit": "^2.0.0",
-    "tslib": "^2.3.1"
-  }
-}
