+++ index.html
@@ -199,10 +199,10 @@
   </template>
 
   <script type="module">
-    import './fixtures/expansion-panel/lib/expansion-panel.js';
-    import './fixtures/fancy-accordion/lib/fancy-accordion.js';
-    import './fixtures/intl-currency/lib/intl-currency.js';
-    import './fixtures/progress-bar/lib/progress-bar.js';
+    import './fixtures/lit/lib/expansion-panel.js';
+    import './fixtures/lit/lib/fancy-accordion.js';
+    import './fixtures/lit/lib/intl-currency.js';
+    import './fixtures/lit/lib/progress-bar.js';
     import './packages/api-viewer/lib/api-viewer.js';
     import './packages/api-docs/lib/api-docs.js';
     import './packages/api-demo/lib/api-demo.js';
