+++ scripts/workspaces-scripts-bin.js
@@ -1,4 +1,5 @@
 import { runWorkspacesScripts } from './runWorkspacesScripts.js';
 
-const script = process.argv[process.argv.length - 1];
-runWorkspacesScripts({ script, concurrency: 5 });
+const [script, folder] = process.argv.slice(2);
+
+runWorkspacesScripts({ script, concurrency: 5, folder });
