+++ workspace-packages.js
@@ -4,10 +4,7 @@ const packages = [
   { name: 'api-docs', type: 'ts', environment: 'browser' },
   { name: 'api-tabs', type: 'ts', environment: 'browser' },
   { name: 'api-viewer', type: 'ts', environment: 'browser' },
-  { name: 'expansion-panel', type: 'ts', environment: 'browser' },
-  { name: 'fancy-accordion', type: 'ts', environment: 'browser' },
-  { name: 'intl-currency', type: 'ts', environment: 'browser' },
-  { name: 'progress-bar', type: 'ts', environment: 'browser' }
+  { name: 'lit', type: 'ts', environment: 'browser' }
 ];
 
 export { packages };
